+++ super-nova-2177/frontend-social-six/app/Loading.jsx
@@ -1,7 +1,0 @@
-export default function Loading() {
-  return (
-    <div className="fixed inset-0 flex items-center justify-center bg-white/50">
-      <img src="/spinner.svg" alt="spinner" />
-    </div>
-  );
-}